% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdapca.R
\name{qdapca}
\alias{qdapca}
\title{Dimension Reduction for Quadratic Discriminant Analysis via Supervised Principal Component Analysis}
\usage{
qdapca(x, y, xnew, rk = 1, include_linear = TRUE, standardize = TRUE)
}
\arguments{
\item{x}{}

\item{y}{}

\item{xnew}{}

\item{rk}{}

\item{include_linear}{}

\item{standardize}{}
}
\value{

}
\description{
This function runs the Dimension Reduction for Quadratic Discriminant Analysis via Supervised Principal Component Analysis (QDAPCA) method.
}
\details{
placeholder
}
\author{
Ruiyang Wu
}
