% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdapca.R
\name{qdapca_cv}
\alias{qdapca_cv}
\title{Fits qdapca using K-fold cross-validation for rank}
\usage{
qdapca_cv(
  x,
  y,
  xnew,
  rk = 1:(min(ncol(x), 10, sqrt(nrow(x)))),
  folds = 5,
  seed = 2020,
  standardize = TRUE
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{xnew}{}

\item{rk}{}

\item{folds}{}

\item{seed}{}

\item{standardize}{}
}
\value{

}
\description{
This functions fits Dimension Reduction for Quadratic Discriminant Analysis via Supervised Principal Component Analysis (QDAPCA) using K-fold cross-validation for rank/dimension
}
\details{
placeholder
}
\author{
Ruiyang Wu
}
